Fix hard failure on google maps error
Fixes: #2175
